Fully implemented TestAddItem test class.
TestAddItem makes two assertions.

testWhenItemIsNull asserts that trying to add a null item to a bill correctly throws IllegalArgumentException.
testWhenItemIsValid adds a valid item to a bill, retrieves it via the getItems method of the bill class, and then asserts that the values of the BillItem's fields are the same.
The amount field of the BillItem is granted an error of 0.001 to account for floating point weirdness.