B: Added Task dataclass and tests.
A Task is a dataclass that holds 3 String fields:
	1. Name, the name of the Task.
	2. Description, a description of the task.
	3. Priority, the priority of the task.

There are simple getters for each field, as well as a setter for the description.
Name and priority are final as they will not be changed past initialization.
The arguments of the constructor (name, description, priority) and of the description setter (description) cannot be null, or they will throw exceptions.
Simple unit tests verify everything working as intended.