2-A:Text field and area for priority and description
Added the text field fxid:selectedPriority which will display the priority of the selected task to the user, along with an appropriate label.
Added the text area fxid:selectedDescription which will display the description of the selected task to the user, along with an appropriate label.

Also moved all the elements around a bit to make it a bit easier for the user to understand and intuit what each button and field is doing.