2-B Made GUI display selected task in read only fields.
The GUI now displays the selected task description and priority.
Also fixed a problem where changing your selection in the tasklist using the keyboard didn't update the description and priority.

I am sure there is a better way to do it than I have, like a specific event listener for when the selected item of a list changes, but for now I have two methods:

	taskSelectedKeyboard and taskSelectedMouse

Both of which perform the same functionality, namely updating the displayed description and priority,
but one listens for a key to be pressed while the other listens for the taskList to be clicked.