4-B: Implemented delete + Major refactoring.
The delete button now works and removes a task from the list.

As I was writing this method, I realized just how hard to read the cod in MainWindow had become, going through layers and layers and layers of JavaFX method calls.
Therefore I refactored a few of the more egregious lines like:

	this.taskList.getSelectionModel().getSelectedItem().setDescription(this.selectedDescription.getText());

into their own methods. In the case of the above snippet, all it's doing is updating the selected item's description, so it is now

	this.updateSelectedDescription();

Much less arcane.