5-C: Connect Logic and UI for Counting Number of Tasks
New method in MainWindow.java: updateNumPriorities, which updates the number of priorities, using the Static method in Utility.java to find the numbers,
then adding them to the label.

It is called inside the addTaskPressed and deleteSelectedTask methods so that any time a task is added or removed (thus the only time a priority could be added or removed)
the priorities get updated accordingly.

Also did some minor cleanup, clearing checkstyle errors and specifically removing the button variable from MainWindow which is not necessary for the button to function.