Added and reprogrammed lab5, minus the tests.
Imported lab 5 starter file.

First updated save student data to save the data in the format described in the instructions (name,grade).
Then updated load student data to load the data in the format described.
Used String.split() to bifurcate each line into a String[] called splitNameAndGrade where splitNameAndGrade[0] is the name and splitNameAndGrade[1] is the String grade.
String grade is then parsed into an integer, and a student is added to the list with the name and grade represented on that line.

Also changed the message for the NumberFormatException of the load method to be a more general "Data was not formatted as name,grade" to better reflect the new state of the program.

It works but it is not tested, so I will implement unit tests in the next commit.