Lab5: Implemented and tested loading student data from CSV file.
loadStudentData now uses Comma-Separated-Value files (CSV) of the form "name,(int)grade."
String.split is used to create an array of strings from each line, separating the strings on either side of commas. These values are used to create Students, which are
then added to an ArrayList, which is converted to an array before finally being returned.