Lab 6: Implemented and tested sorting
The sorting combo box now displays two new options:

	Ascending Order
	Descending Order

When selected from the box, the list of tasks will now sort the tasklist by priority in the selected order.
This is achieved through the use of two new custom comparators housed within the combo box, which are defined in Ascending.java and Descending.java.

There are new JUnit tests in place to ensure that the compare methods inside Ascending and Descending work as intended.

Ascending and descending both Override Object:toString and return the strings "Ascending Order" and "Descending Order" respectively.