2.A/B: Added the ability to sort based on name.
Refactored original Ascending and Descending comparators to AscendingPriority and DescendingPriority.
Added AscendingName and DescendingName comparators and unit tests for both.
Added them to the combo box of comparators allowing for them to be used as Ascending and Descending have been in the past.

They use the standard java String.compareTo method to sort tasks in ascending and descending lexicographical order (Alphabetical order including numbers symbols and special chars).