3.A: Added support for subtasks in MainWindow.
Application now has a button and listview whose sole concern will be supporting subtasks. Associated methods and field stubs added to MainWindow.
Currently does nothing, as the underlying logic doesn't yet support subtasks, but it should be easy to get it working from this state once that functionality is implemented.