3.B: Implemented Business Logic for Sub-Tasks.
There is now underlying logic for subtasks.

The ContainerTask class inherits from task and contains (as the name suggests) a list of tasks. It is used in situations where there are subtasks for a given task.
Task has new methods to support the inheritance relationship.

addTask adds a task to a ContainerTask and returns it. When called on a ContainerTask it returns itself.
getSubTasks returns a list of all the subTasks of a given ContainerTask. A Task, having no subtasks will always return an empty list.