3.C: Connected UI to Business Logic for Subtasks
The UI now allows for the adding of subtasks to a task.

To do so, you simply select a task, input a name, description and priority in the same fields you would for a normal task and click the "Add Sub-Task" button.
If you try to add a Sub-Task without a task selected, the program will tell you to select one.
The Sub-Task list updates whenever a sub-task is added, and your selected task will stay selected even if you click off so you can add as many subtasks as you want
to a selected task without having to re-select it each time.