Lab 7: Integrated Viewmodel with Code Behind.
The program is functionally identical, but now all of the logic that was previously handled by the MainWindow class, is now handled by the GeneratorViewModel class.
The only thing MainWindow handles now are UI concerns, like showing alerts when exceptions are thrown, or listening for the button to be pressed.