Project 3: Added and refactored base project.
I am working from the JavaFX sample starter.
Lots of refactoring to ensure that the names make sense and all references to what it used to be are eliminated.
Also changed the starting VBox in the fxml to an AnchorPane so I can rearrange elements more easily.

The deletions are a previous attempt at this.